{{cookiecutter.directory_name}}/pyproject.toml
[tool.poetry]
name = "Boilerplate"
version = "0.1.0"
description = "Boilerplate ebs"
authors = ["John Proud"]

[tool.poetry.dependencies]
python = "3.8"
Django = "^3.1.2"
django-environ = "^0.4.5"
djangorestframework = "^3.12.1"
gunicorn = "^20.0.4"
django-cors-headers = "^3.5.0"
djangorestframework-simplejwt = "^4.4.0"
drf-yasg = "^1.17.1"
psycopg2 = "^2.8.6"
raven = "^6.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
